+++ circuit-retained/src/androidInstrumentedTest/kotlin/com/slack/circuit/retained/android/RetainedSaveableTest.kt
@@ -49,7 +49,8 @@ class RetainedSaveableTest {
     var continuity: ContinuityViewModel? = null
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-      @Suppress("UNCHECKED_CAST") return ContinuityViewModel().also { continuity = it } as T
+      @Suppress("UNCHECKED_CAST")
+      return ContinuityViewModel().also { continuity = it } as T
     }
   }