+++ circuit-test/src/commonMain/kotlin/com/slack/circuit/test/PresenterTestExtensions.kt
@@ -10,12 +10,17 @@ import app.cash.turbine.test
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.presenter.Presenter
 import kotlin.time.Duration
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.distinctUntilChanged
 
 /**
  * Presents this [Presenter] and invokes a `suspend` [ReceiveTurbine] [block] that can be used to
  * assert state emissions from this presenter.
  *
  * @param timeout an optional timeout for the test. Defaults to 1 second (in Turbine) if undefined.
+ * @param moleculeFlowTransformer an optional transformer for the underlying [moleculeFlow]. Must
+ *   still return a [Flow] of type [UiState], but can be used for custom filtering. By default, it
+ *   runs [distinctUntilChanged].
  * @param block the block to invoke.
  * @see moleculeFlow
  * @see test
@@ -23,9 +28,10 @@ import kotlin.time.Duration
 public suspend fun <UiState : CircuitUiState> Presenter<UiState>.test(
   timeout: Duration? = null,
   name: String? = null,
+  moleculeFlowTransformer: (Flow<UiState>) -> Flow<UiState> = Flow<UiState>::distinctUntilChanged,
   block: suspend ReceiveTurbine<UiState>.() -> Unit,
 ) {
-  presenterTestOf({ present() }, timeout, name, block)
+  presenterTestOf({ present() }, timeout, name, moleculeFlowTransformer, block)
 }
 
 /**
@@ -34,6 +40,9 @@ public suspend fun <UiState : CircuitUiState> Presenter<UiState>.test(
  *
  * @param presentFunction the [Composable] present function being tested.
  * @param timeout an optional timeout for the test. Defaults to 1 second (in Turbine) if undefined.
+ * @param moleculeFlowTransformer an optional transformer for the underlying [moleculeFlow]. Must
+ *   still return a [Flow] of type [UiState], but can be used for custom filtering. By default, it
+ *   runs [distinctUntilChanged].
  * @param block the block to invoke.
  * @see moleculeFlow
  * @see test
@@ -42,7 +51,10 @@ public suspend fun <UiState : CircuitUiState> presenterTestOf(
   presentFunction: @Composable () -> UiState,
   timeout: Duration? = null,
   name: String? = null,
+  moleculeFlowTransformer: (Flow<UiState>) -> Flow<UiState> = Flow<UiState>::distinctUntilChanged,
   block: suspend ReceiveTurbine<UiState>.() -> Unit,
 ) {
-  moleculeFlow(RecompositionMode.Immediate, presentFunction).test(timeout, name, block)
+  moleculeFlow(RecompositionMode.Immediate, presentFunction)
+    .run(moleculeFlowTransformer)
+    .test(timeout, name, block)
 }